requirements.txt
aiohttp==3.8.4
asyncio==3.4.3
pyyaml==6.0
psycopg2-binary==2.9.6
schedule==1.2.0 